removing layer that saves to sfmc
